Process to find Fraud Detection using ABN
Process to find Fraud Detection using ABN. The bot will look for the ABN details in the ABN lookup website and gets all the details such as ABN status, GST registered and Vendor Name, Address and a matching Google Address. It then matches the Google Address with the list of employees which is a separate list and if a match is found it displays the vendor name.